Removed some troubleshooting code
Removed a comment and a line of code regarding finding the column names of the initial CSV file in the JHU repository; the CSV files contained in the repository had two sets of column variable nomenclature upon further inspection (Total_Test_Results was changed at one point to People_Tested, and Case_Fatality_Ratio was changed to Mortality_Rate). The code was unnecessary after discovering the discrepancy and accounting for it in the data import.